% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stellarscope_report.R
\name{load_labeled_mtx}
\alias{load_labeled_mtx}
\title{Load a sparse matrix with row and column labels}
\usage{
load_labeled_mtx(mtx_file, rowname_file, colname_file)
}
\arguments{
\item{mtx_file}{Sparse matrix file in MTX format}

\item{rowname_file}{TSV file with row labels in first column.}

\item{colname_file}{TSV file with column labels in first column.}
}
\value{
A sparse matrix of of class "dgTMatrix" with row and columns labels.
}
\description{
Sparse matrixes in the MTX format have integer indexes and do not have
row or column names included. The labels may be included in separate
files. This function takes three input files and returns a labeled sparse
matrix. Dimensions of the returned sparse matrix are match the provided row
and column names, transposing the matrix if needed.
}
