% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tsv_counts.R
\name{load_star_counts}
\alias{load_star_counts}
\title{Load counts from multiple STAR tab-delimited reports}
\usage{
load_star_counts(
  files,
  colnames = names(files),
  stranded = c("unstranded", "read1", "read2")
)
}
\arguments{
\item{files}{A character vector with paths to the reports.}

\item{colnames}{Output column names. If \code{files} is a named vector, the names
will be used as column names. Otherwise finds a unique string
from the report filenames.}

\item{stranded}{Strandedness of data. Use 'unstranded' for unstranded data,
'read1' if the first read aligns to the RNA, and 'read2' if the second
read aligns to the RNA. dUTP-based methods should use 'read2'.}
}
\value{
A dataframe where the rows are genes/loci and the columns are
samples (when each report corresponds to one sample).
}
\description{
This returns a dataframe where the rows are genes/loci and columns are
samples. If \code{files} is a named vector, the names are used as sample
names; otherwise sample names can be provided by \code{colnames}. If not
provided, unique strings will be extracted from the report filenames.
}
\examples{
ddir <- system.file("extdata", package="scopetools")
tsv_files <- Sys.glob(file.path(ddir, '*.ReadsPerGene.out.tab'))

gene_counts <- load_star_counts(tsv_files)
}
