% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stellarscope_report.R
\name{load_stellarscope_report}
\alias{load_stellarscope_report}
\title{Load count matrix from Stellarscope output files}
\usage{
load_stellarscope_report(
  stellarscope_dir,
  exp_tag = "stellarscope",
  count_filename = "*TE_counts.mtx",
  feature_filename = "*features.tsv",
  barcode_filename = "*barcodes.tsv"
)
}
\arguments{
\item{stellarscope_dir}{Stellarscope output directory.}

\item{exp_tag}{Experiment tag used in stellarscope run.
Default is "stellarscope".}

\item{count_filename}{File name pattern for count matrix (MTX format). May
contain wildcards.
Default is '*TE_counts.mtx'.}

\item{feature_filename}{File name pattern for feature list (tab-separated,
features in first column). May contain wildcards.
Default is '*features.tsv'.}

\item{barcode_filename}{File name pattern for barcode list (tab-separated,
barcodes in first column). May contain wildcards.
Default is '*barcodes.tsv'.}
}
\value{
A sparse matrix of of class "dgTMatrix" with features as rows and
cells as columns.
}
\description{
Load count matrix from Stellarscope output files
}
