% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stellarscope_seurat.R
\name{load_stellarscope_seurat}
\alias{load_stellarscope_seurat}
\title{Load count matrix from Stellarscope output files}
\usage{
load_stellarscope_seurat(
  stellarscope_dir,
  starsolo_dir = NULL,
  project = "StellarscopeProject",
  min.cells = 0,
  min.features = 0,
  TE_count_file = Sys.glob(file.path(stellarscope_dir, "*TE_counts.mtx")),
  TE_feature_file = Sys.glob(file.path(stellarscope_dir, "*features.tsv")),
  TE_barcode_file = Sys.glob(file.path(stellarscope_dir, "*barcodes.tsv")),
  GENE_count_file = Sys.glob(file.path(starsolo_dir, "*matrix.mtx")),
  GENE_feature_file = Sys.glob(file.path(starsolo_dir, "*features.tsv")),
  GENE_barcode_file = Sys.glob(file.path(starsolo_dir, "*barcodes.tsv")),
  TE_metadata = data.frame(retro.hg38.v1, row.names = "locus"),
  GE_metadata = NULL
)
}
\arguments{
\item{stellarscope_dir}{Stellarscope output directory}

\item{starsolo_dir}{STARsolo output directory}

\item{project}{Project name for the \code{Seurat} object}

\item{min.cells}{Include features detected in at least this many cells.
Will subset the counts matrix as well. To reintroduce excluded features,
create a new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many features are
detected.}

\item{TE_count_file}{File name for Stellarscope count matrix in MTX format.
Default is a file matching '*TE_counts.mtx' in \code{stellarscope_dir}.}

\item{TE_feature_file}{File name for Stellarscope feature list in TSV
format. Default is a file matching '*features.tsv' in \code{stellarscope_dir}.}

\item{TE_barcode_file}{File name for Stellarscope barcode list in TSV
format. Default is a file matching '*barcodes.tsv' in \code{stellarscope_dir}.}

\item{GENE_count_file}{File name for STARsolo count matrix in MTX format.
Default is a file matching '*matrix.mtx' in \code{starsolo_dir}.}

\item{GENE_feature_file}{File name for STARsolo feature list in TSV format.
Default is a file matching '*features.tsv' in \code{starsolo_dir}.}

\item{GENE_barcode_file}{File name for STARsolo barcode list in TSV format.
Default is a file matching '*barcodes.tsv' in \code{starsolo_dir}.}

\item{TE_metadata}{Data frame with metadata related to the TE features}

\item{GE_metadata}{Data frame with metadata related to the gene features}
}
\value{
A \code{Seurat} object with counts from stellarscope and STARsolo.
}
\description{
This returns a \code{Seurat} object with raw counts from Stellarscope and
STARsolo (if provided). Only cell barcodes present in both are
retained. Feature metadata can also be included for TEs and genes.
}
