% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_telescope_reports.R
\name{load_telescope_reports}
\alias{load_telescope_reports}
\title{Load counts from multiple Telescope reports}
\usage{
load_telescope_reports(
  files,
  colnames = names(files),
  all_locs = NULL,
  count_column = "final_count"
)
}
\arguments{
\item{files}{A character vector with paths to the reports.}

\item{colnames}{Output column names. If \code{files} is a named vector, the names
will be used as column names. Otherwise finds a unique string
from the report filenames.}

\item{all_locs}{A character vector with all genes/loci to be included
in the output; typically this is a list of all loci in the annotation.
If not provided, the loci will be the union of loci in all reports, in
an arbitrary order.}

\item{count_column}{The column in the Telescope report to use.
Defaults is 'final_count'}
}
\value{
A dataframe where the rows are genes/loci and the columns are
samples (when each report corresponds to one sample).
}
\description{
This returns a dataframe where the rows are genes/loci and columns are
samples. If \code{files} is a named vector, the names are used as sample
names; otherwise sample names can be provided by \code{colnames}. If not
provided, unique strings will be extracted from the report filenames.
}
\examples{
ddir <- system.file("extdata", package="scopetools")
t_files <- file.path(ddir, paste0(samples.encode$GEO_sample, '-telescope_report.tsv'))
names(t_files) <- samples.encode$Name

# Provide all_locs to include all annotated loci (preferred)
herv_counts <- load_telescope_reports(t_files, all_locs=HERVrmsk.hg38.v2$locus)

# Contstruct table with loci that appear in reports
\dontrun{
herv_counts <- load_telescope_reports(t_files)
}

}
