% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tsv_counts.R
\name{load_tsv_counts}
\alias{load_tsv_counts}
\title{Load counts from multiple tab-delimited reports}
\usage{
load_tsv_counts(
  files,
  colnames = names(files),
  all_locs = NULL,
  header = FALSE,
  gene_id_column = 1,
  count_column = 2,
  exclude_features = c()
)
}
\arguments{
\item{files}{A character vector with paths to the reports.}

\item{colnames}{Output column names. If \code{files} is a named vector, the names
will be used as column names. Otherwise finds a unique string
from the report filenames.}

\item{all_locs}{A character vector with all genes/loci to be included
in the output; typically this is a list of all loci in the annotation.
If not provided, the loci will be the union of loci in all reports, in
an arbitrary order.}

\item{header}{Whether the first line of the input files is a header.}

\item{gene_id_column}{Report column number containing the gene ID.}

\item{count_column}{Report column number containing the count.}

\item{exclude_features}{Gene/feature IDs to be excluded from dataframe.}
}
\value{
A dataframe where the rows are genes/loci and the columns are
samples (when each report corresponds to one sample).
}
\description{
This returns a dataframe where the rows are genes/loci and columns are
samples. If \code{files} is a named vector, the names are used as sample
names; otherwise sample names can be provided by \code{colnames}. If not
provided, unique strings will be extracted from the report filenames.
}
\examples{
ddir <- system.file("extdata", package="scopetools")
tsv_files <- Sys.glob(file.path(ddir, '*.ReadsPerGene.out.tab'))

gene_counts <- load_tsv_counts(tsv_files)

}
