% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stellarscope_cell_qc.R
\name{stellarscope_cell_qc}
\alias{stellarscope_cell_qc}
\title{Filter cells from Seurat object}
\usage{
stellarscope_cell_qc(object)
}
\arguments{
\item{object}{Seurat object (with TE feature metadata)}
}
\value{
Seurat object with cells passing QC
}
\description{
This function expects that the Seurat object includes TE feature metadata.
Specifically, \code{feattype} should be set to "TE" for TE features and \code{te_class}
should indicate whether the TE is an "LTR" or "LINE" retroelement. Seurat
objects created with \code{\link[=load_stellarscope_seurat]{load_stellarscope_seurat()}} will include
the necessary metadata.
}
\details{
The function adds metadata for the mitochondrial percentage (\code{percent.mt}),
HERV percentage (\code{percent.HERV}), L1 percentage (\code{percent.L1}), and total TE
percentage (\code{percent.TE}). Outliers are detected based on the median absolute
deviation (MAD) implemented in \code{\link[scuttle:isOutlier]{scuttle::isOutlier()}}. Cells are excluded
based on extreme RNA count, extreme feature count, or high MT percentage.
}
\examples{
# Seurat object
pbmc <- stellarscope_cell_qc(pbmc.orig)

}
