% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_names_from_file_list}
\alias{unique_names_from_file_list}
\title{Find unique names from list of files}
\usage{
unique_names_from_file_list(files)
}
\arguments{
\item{files}{A character vector with filenames}
}
\description{
Find unique names from list of files
}
