% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stellarscope_report.R
\name{load_stellarscope_report}
\alias{load_stellarscope_report}
\title{Load count matrix from Stellarscope output files}
\usage{
load_stellarscope_report(
  dirname,
  count_file_suffix = "TE_counts.mtx",
  feature_file_suffix = "features.tsv",
  barcode_file_suffix = "barcodes.tsv"
)
}
\arguments{
\item{dirname}{Stellarscope output directory.}

\item{count_file_suffix}{File name suffix for count matrix in MTX format. Default is 'TE_counts.mtx'.}

\item{feature_file_suffix}{File name suffix for feature list. Default is 'features.tsv'.}

\item{barcode_file_suffix}{File name suffix for barcode list. Default is 'barcodes.tsv'.}
}
\value{
A sparse matrix of of class "dgTMatrix" with features as rows and cells as columns.
}
\description{
Load count matrix from Stellarscope output files
}
