% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stellarscope_seurat.R
\name{load_stellarscope_seurat}
\alias{load_stellarscope_seurat}
\title{Load count matrix from Stellarscope output files}
\usage{
load_stellarscope_seurat(
  stellarscope_dir,
  starsolo_dir = NULL,
  exp_tag = "stellarscope",
  project = "StellarscopeProject",
  min.cells = 0,
  min.features = 0,
  remove.nofeat = TRUE,
  TE_count_filename = "*TE_counts.mtx",
  TE_feature_filename = "*features.tsv",
  TE_barcode_filename = "*barcodes.tsv",
  GENE_count_filename = "*matrix.mtx",
  GENE_feature_filename = "*features.tsv",
  GENE_barcode_filename = "*barcodes.tsv",
  TE_metadata = data.frame(retro.hg38.v1, row.names = "locus"),
  GE_metadata = NULL
)
}
\arguments{
\item{stellarscope_dir}{Stellarscope output directory}

\item{starsolo_dir}{STARsolo output directory, optional.}

\item{exp_tag}{Experiment tag used in stellarscope run.
Default is "stellarscope".}

\item{project}{Project name for the \code{Seurat} object}

\item{min.cells}{Include features detected in at least this many cells.
Will subset the counts matrix as well. To reintroduce excluded features,
create a new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many features are
detected.}

\item{remove.nofeat}{Remove "no feature" category from Stellarscope counts.
Default is TRUE.}

\item{TE_count_filename}{File name pattern for TE count matrix (MTX format).
May contain wildcards.Default is '*TE_counts.mtx'.}

\item{TE_feature_filename}{File name pattern for TE feature list
(tab-separated, features in first column). May contain wildcards.
Default is '*features.tsv'.}

\item{TE_barcode_filename}{File name pattern for TE barcode list
(tab-separated, barcodes in first column). May contain wildcards.
Default is '*barcodes.tsv'.}

\item{GENE_count_filename}{File name pattern for gene count matrix
(MTX format). May contain wildcards. Default is '*matrix.mtx'.}

\item{GENE_feature_filename}{File name pattern for gene feature list
(tab-separated, features in first column). May contain wildcards.
Default is '*features.tsv'.}

\item{GENE_barcode_filename}{File name pattern for gene barcode list
(tab-separated, barcodes in first column). May contain wildcards.
Default is '*barcodes.tsv'.}

\item{TE_metadata}{Data frame with metadata related to the TE features}

\item{GE_metadata}{Data frame with metadata related to the gene features}
}
\value{
A \code{Seurat} object with counts from stellarscope and STARsolo.
}
\description{
This returns a \code{Seurat} object with raw counts from Stellarscope and
STARsolo (if provided). Only cell barcodes present in both are
retained. Feature metadata can also be included for TEs and genes.
}
