% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_telescope_reports.R
\name{load_telescope_reports}
\alias{load_telescope_reports}
\title{Load counts from multiple Telescope reports}
\usage{
load_telescope_reports(files, count_column = "final_count", all_locs = NULL)
}
\arguments{
\item{files}{A character vector with paths to the reports. If a named vector, the
names will be used as the column names.}

\item{count_column}{The column in the Telescope report to use.
Defaults is 'final_count'}

\item{all_locs}{A character vector with the loci to include, possibly
extracted from the annotation. If not provided, all loci appearing
in the reports will be used in an arbitrary order.}
}
\value{
A dataframe where the rows are genes/loci and the columns are
reports. Typically each report will correspond to one sample.
}
\description{
This returns a dataframe where the rows are genes/loci with one column per
report. (Typically each report corresponds to one sample). The  and the columns are
}
\examples{
\dontrun{
   t_files <- file.path('samples', samples$sample_id, 'telescope.report.tsv')
   names(t_files) <- samples$sample_id
   herv_counts <- load_telescope_reports(t_files)
   
   # If you have the annotation locs in a table, this will maintain
   # the ordering
   herv_counts <- load_telescope_reports(t_files, retro.annot$locus)
   
}
}
