% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tsv_counts.R
\name{load_tsv_counts}
\alias{load_tsv_counts}
\title{Load counts from multiple tab-delimited reports}
\usage{
load_tsv_counts(
  files,
  colnames = names(files),
  all_locs = NULL,
  count_column = "final_count"
)
}
\description{
This returns a dataframe where the rows are genes/loci and columns are
samples. If \code{files} is a named vector, the names are used as sample
names; otherwise sample names can be provided by \code{colnames}. If not
provided, unique strings will be extracted from the report filenames.
}
